+++ src/main/kotlin/com/coder/toolbox/views/AuthWizardPage.kt
@@ -21,11 +21,11 @@ class AuthWizardPage(
     private val settingsPage: CoderSettingsPage,
     private val visibilityState: MutableStateFlow<ProviderVisibilityState>,
     initialAutoLogin: Boolean = false,
-    onConnect: (
+    onConnect: suspend (
         client: CoderRestClient,
         cli: CoderCLIManager,
     ) -> Unit,
-) : CoderPage(context, context.i18n.ptrl("Authenticate to Coder"), false) {
+) : CoderPage(context.i18n.ptrl("Authenticate to Coder"), false) {
     private val shouldAutoLogin = MutableStateFlow(initialAutoLogin)
     private val settingsAction = Action(context.i18n.ptrl("Settings"), actionBlock = {
         context.ui.showUiPage(settingsPage)