+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Fixed
 
-- outdated and stopped workspaces are now updated and started when handling URI 
+- `Stop` action is now available for running workspaces that have an out of date template.
+- outdated and stopped workspaces are now updated and started when handling URI
 
 ## 0.3.0 - 2025-06-10